+++ tests/assets/sample_tasks/sample.R
@@ -1,2 +1,5 @@
 # + tags=["parameters"]
 c(1, 2, 3)
+
+# +
+42