+++ tests/assets/sample_tasks/sample.py
@@ -1,2 +1,5 @@
 # + tags=["parameters"]
 1 + 1
+
+# +
+42