+++ tests/cli/test_install.py
@@ -11,7 +11,7 @@ import pytest
 from click.testing import CliRunner
 
 from ploomber.cli import install as install_module
-from ploomber.cli.cli import install
+from ploomber_cli.cli import install
 from ploomber.cli.install import _pip_install
 from ploomber.exceptions import BaseException
 from conftest import (_write_sample_conda_env, _prepare_files,