+++ tests/spec/test_dagspec.py
@@ -1064,6 +1064,9 @@ def test_import_tasks_from_with_non_empty_env(tmp_nbs):
 # + tags=["parameters"]
 some_param = None
 # -
+
+# +
+1 + 1
 """)
     spec_d = yaml.safe_load(Path('pipeline.yaml').read_text())
     spec_d['meta']['import_tasks_from'] = 'some_tasks.yaml'
@@ -1457,11 +1460,17 @@ def test_grid_and_upstream_wildcard_scripts(spec, tmp_directory):
 # + tags=['parameters']
 upstream = None
 param = None
+
+# +
+1 + 1
 """)
 
     Path('downstream.py').write_text("""
 # + tags=['parameters']
 upstream = ['upstream-*']
+
+# +
+1 + 1
 """)
 
     spec = DAGSpec(spec)
@@ -1550,11 +1559,17 @@ def test_spec_from_yaml_resolves_paths_from_wildcard(tmp_directory, spec):
 # + tags=['parameters']
 upstream = None
 param = None
+
+# +
+1 + 1
 """)
 
     Path('downstream.py').write_text("""
 # + tags=['parameters']
 upstream = ['upstream-*']
+
+# +
+1 + 1
 """)
 
     spec = DAGSpec(spec)
@@ -1895,3 +1910,41 @@ tasks:
     assert "script.py is missing the parameters cell" in captured.out
     assert cell
     assert idx == 0
+
+
+@pytest.mark.parametrize('home', [
+    'dir',
+    'some/nested/dir',
+])
+def test_dagspec_infers_name(tmp_directory, home):
+    home = Path(home)
+    home.mkdir(parents=True)
+
+    path = home / 'script.py'
+    path.touch()
+
+    (home / 'pipeline.yaml').write_text("""
+tasks:
+    - source: script.py
+      product: output.ipynb
+""")
+
+    dag = DAGSpec(home / 'pipeline.yaml').to_dag()
+
+    assert dag.name == 'dir'
+
+
+def test_dagspec_from_dir_doesnt_assign_name(tmp_directory):
+    home = Path('dir')
+    home.mkdir()
+    path = home / 'script.py'
+    path.touch()
+
+    dag = DAGSpec({
+        'tasks': [{
+            'source': 'dir/script.py',
+            'product': 'output.ipynb'
+        }]
+    }).to_dag()
+
+    assert dag.name == 'No name'